rafactor complex, cossin, atan2
