+++ dsp/src/lib.rs
@@ -2,10 +2,6 @@
 #![cfg_attr(feature = "nightly", feature(asm, core_intrinsics))]
 
 use core::ops::{Add, Mul, Neg};
-use serde::{Deserialize, Serialize};
-
-#[derive(Copy, Clone, Default, Deserialize, Serialize)]
-pub struct Complex<T>(pub T, pub T);
 
 /// Bit shift, round up half.
 ///
@@ -116,13 +112,19 @@ where
         .fold(y0, |y, xa| y + xa)
 }
 
+mod atan2;
+mod complex;
+mod cossin;
 pub mod iir;
 pub mod iir_int;
 pub mod lockin;
 pub mod pll;
 pub mod reciprocal_pll;
-pub mod trig;
 pub mod unwrap;
+
+pub use atan2::atan2;
+pub use complex::Complex;
+pub use cossin::cossin;
 
 #[cfg(test)]
 pub mod testing;
